+++ database/migrations/2025_05_20_081407_create_asignaturas_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('asignaturas', function (Blueprint $table) {
-            $table->integer('idasignaturas')->primary();
+            $table->unsignedBigInteger('idasignaturas')->primary();
             $table->string('nombre_asignatura');
             $table->timestamps();
         });
